fast_api/requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.5.0
asttokens==2.4.1
async-timeout==4.0.3
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
dataclasses-json==0.6.7
decorator==5.1.1
distro==1.9.0
exceptiongroup==1.2.2
executing==2.1.0
fastapi==0.115.0
frozenlist==1.4.1
greenlet==3.1.1
groq==0.11.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.10
ipython==8.18.1
jedi==0.19.1
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.0
langchain-community==0.3.0
langchain-core==0.3.5
langchain-groq==0.2.0
langchain-text-splitters==0.3.0
langgraph==0.2.23
langgraph-checkpoint==1.0.10
langsmith==0.1.125
marshmallow==3.22.0
matplotlib-inline==0.1.7
msgpack==1.1.0
multidict==6.1.0
mypy-extensions==1.0.0
numpy==1.26.4
orjson==3.10.7
packaging==24.1
parso==0.8.4
pexpect==4.9.0
prompt-toolkit==3.0.47
ptyprocess==0.7.0
pure-eval==0.2.3
pydantic==2.9.2
pydantic-core==2.23.4
pydantic-settings==2.5.2
pygments==2.18.0
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.35
stack-data==0.6.3
starlette==0.38.5
tenacity==8.5.0
traitlets==5.14.3
typing-extensions==4.12.2
typing-inspect==0.9.0
urllib3==2.2.3
uvicorn==0.30.6
wcwidth==0.2.13
yarl==1.11.1
